test(server): split transport contract owners
